# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Dopamine Market API
  description: >
    Синтетический маркетплейс/сервис доставки — портфолио-проект.
    Монетизация: фиксированная плата 10₽ за оформленный заказ (INV-01).
    Полностью синтетический каталог и курьеры, симуляция доставки —
    таймер-стейт-машина (см. docs/prd/PRD.md, ADR-005, ADR-009).
  version: "1.0.0"
  contact:
    name: Dopamine Market (портфолио-проект)
servers:
  - url: http://localhost:8080
    description: Локальная разработка
  - url: https://api.dopamine-market.example.com
    description: Продакшн (заглушка домена)
...
paths:
...
  /auth/register:
    post:
      tags: [identity]
      summary: Регистрация нового пользователя
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: "strongPassword123"
      responses:
        "201":
          description: Успешная регистрация, токены в httpOnly cookies и теле ответа
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  access_token:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Email уже занят
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/login:
    post:
      tags: [identity]
      summary: Аутентификация пользователя
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  format: password
                  example: "strongPassword123"
      responses:
        "200":
          description: Успешный вход
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  access_token:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/quick-login:
    post:
      tags: [identity]
      summary: Быстрая аутентификация в 1 клик (Google, Apple, Yandex, Quick Email)
      description: Создает или находит пользователя по email/провайдеру и сразу выдает сессионные куки и access_token.
      operationId: quickLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  example: "user@gmail.com"
                nickname:
                  type: string
                  example: "CyberDopamine"
                provider:
                  type: string
                  enum: [google, yandex, apple, quick_email]
                  example: "google"
      responses:
        "200":
          description: Успешная быстрая авторизация
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  access_token:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/refresh:
    post:
      tags: [identity]
      summary: Ротация пары токенов
      operationId: refreshTokens
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
      responses:
        "200":
          description: Токены обновлены
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /cart:
    get:
      tags: [cart]
      summary: Просмотр корзины с актуальными данными товаров и ПВЗ
      operationId: getCart
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      tags: [cart]
      summary: Очистка корзины
      operationId: clearCart
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
  schemas:
    User:
      type: object
      required: [id, email, created_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
...
    PickupPoint:
      type: object
      required: [id, name, latitude, longitude, distance_meters, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        distance_meters:
          type: number
          format: double
        created_at:
          type: string
          format: date-time
    CartItem:
      type: object
      required: [product_id, category_id, name, price_rub, quantity, subtotal_rub]
      properties:
        product_id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
        category_name:
          type: string
        name:
          type: string
        price_rub:
          type: string
          example: "10.00"
        image_seed:
          type: string
        quantity:
          type: integer
        subtotal_rub:
          type: string
          example: "20.00"
    Cart:
      type: object
      required: [items, total_quantity, total_price_rub, fixed_order_cost]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        pickup_point:
          $ref: "#/components/schemas/PickupPoint"
        total_quantity:
          type: integer
        total_price_rub:
          type: string
          example: "20.00"
        fixed_order_cost:
          type: string
          example: "10.00"
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    TooManyRequests:
      description: Превышен лимит запросов
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"